Added Omega Files
The print.css and style.css files need to replace the files of the same
name in omeka/themes/emiglio/css/

The show.php file needs to replace the file of the same name in
omeka/themes/emiglio/items/

The record-metadata.php file needs to replace the file of the same name
in omeka/applications/views/scripts/common/

The CorrectionsPlugin.php file updates the same file in
omeka/plugins/Corrections/ :I added a little hack so that the
Corrections plugin link can be styled separately from other links in
Emiglio.

Here is a list of the changes I made to the files: for the css files I
changed the fonts & colors to reflect the fonts & colors used on the
new library website; in the show.php I added a second call for the
corrections plugin and placed it in a secondary div; for the
record-metadata.php I commented out an <h1> so that the words 'Dublin
Core' won't show on the item pages.